TutorialFileReset
Teacher loaded a different file, changing the format and titles.
What is in the pastebin is separate from what the last code shown in two videos in a row.
Started a note on what to learn from what I can't understand (what's declaring the locations, how are the functions interacting, why sequenced in what way)
Teacher got confused and anxious and then that carried over to the explanations.
TLDR: Tutorial dude had to restart the project so I am too